// frontend/manager-web/src/pages/calls/components/grid/types.d.ts
export interface IRow {
  id: number;
  callId: number;
  callType: string;
  startAt: string;
  leadName: string;
  userName: string;
  brand: string;
  group: string;
  salesStatusAfter: string;
  duration: string;
  billDuration: string;
  hangup: string;
  callerId: string;
  phoneDestination: string;
}